Pitch shifter processing class (#2)
* Refactor delay interpolation modes, and add pitch shifter class

* Pitch Shift: fix bugs and adjust default settings

* Pitch shift: add unit tests and documentation

* Fix warnings

* Fix pitch-shift unit test bugs

+++ DSP/Filters/chowdsp_IIRFilter.h
@@ -7,7 +7,7 @@ namespace chowdsp
  * Uses Transposed Direct Form II:
  * https://ccrma.stanford.edu/~jos/fp/Transposed_Direct_Forms.html
  */
-template<int order, typename FloatType=float>
+template<size_t order, typename FloatType=float>
 class IIRFilter
 {
 public:
